requirements.txt
fastapi
uvicorn[standard]
streamlit
pydantic
python-dotenv
sentence-transformers
faiss-cpu
PyMuPDF
python-docx
requests
ruff
pytest
google-generativeai
python-multipart